+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Add tests for namespaced XML (#292)
 - Always use Composer-installed versions of the dev tools (#291)
 - Fix warnings in the `travis.yml` (#289)
 - Do not cache `vendor/` on Travis CI (#288)